basic authorization

+++ app/Http/Controllers/IdeaController.php
@@ -14,12 +14,19 @@ class IdeaController extends Controller
             'content' => 'required|min:3|max:240'
         ]);
 
+        $validated['user_id'] = auth()->user()->id;
+
         Idea::create($validated);
 
         return redirect()->route('dashboard')->with('success','Idea created successfully!');
     }
 
     public function destroy(Idea $idea) {
+
+        if(auth()->user()->id !== $idea->user_id) {
+            abort(404);
+        }
+
         $idea->delete();
 
         return redirect()->route('dashboard')->with('success','Idea deleted successfully!');
@@ -30,11 +37,17 @@ class IdeaController extends Controller
     }
 
     public function edit(Idea $idea) {
+        if(auth()->user()->id !== $idea->user_id) {
+            abort(404);
+        }
         $editing = true;
         return view('ideas.show', compact('idea', "editing"));
     }
 
     public function update(Idea $idea) {
+        if(auth()->user()->id !== $idea->user_id) {
+            abort(404);
+        }
         $validated = request()->validate([
             'content' => 'required|min:3|max:240'
         ]);
